+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -2,8 +2,8 @@ package com.techcourse;
 
 import jakarta.servlet.ServletContext;
 import nextstep.mvc.DispatcherServlet;
-import nextstep.mvc.controller.RequestMappingHandlerAdapter;
-import nextstep.mvc.controller.RequestMappingHandlerMapping;
+import nextstep.mvc.adapter.RequestMappingHandlerAdapter;
+import nextstep.mvc.mapping.RequestMappingHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
