+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -4,10 +4,10 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.HandlerAdapter;
-import nextstep.mvc.controller.HandlerAdapterRegistry;
-import nextstep.mvc.controller.HandlerMapping;
-import nextstep.mvc.controller.HandlerMappingRegistry;
+import nextstep.mvc.adapter.HandlerAdapter;
+import nextstep.mvc.adapter.HandlerAdapterRegistry;
+import nextstep.mvc.mapping.HandlerMapping;
+import nextstep.mvc.mapping.HandlerMappingRegistry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
