+++ mvc/src/test/java/nextstep/mvc/RequestMappingHandlerAdapterTest.java
@@ -3,8 +3,8 @@ package nextstep.mvc;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.stream.Stream;
-import nextstep.mvc.controller.HandlerExecution;
-import nextstep.mvc.controller.RequestMappingHandlerAdapter;
+import nextstep.mvc.adapter.RequestMappingHandlerAdapter;
+import nextstep.mvc.mapping.HandlerExecution;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
