+++ mvc/src/test/java/nextstep/mvc/controller/HandlerAdapterRegistryTest.java
@@ -5,6 +5,12 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.Map;
+import nextstep.mvc.adapter.HandlerAdapter;
+import nextstep.mvc.adapter.HandlerAdapterRegistry;
+import nextstep.mvc.adapter.RequestMappingHandlerAdapter;
+import nextstep.mvc.mapping.ControllerScanner;
+import nextstep.mvc.mapping.HandlerExecution;
+import nextstep.mvc.mapping.HandlerKey;
 import nextstep.web.support.RequestMethod;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
